requirements.txt
Flask
MongoEngine
docker
python-dotenv
openai